requirements.txt
pybit==5.6.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
schedule==1.2.1
flask==3.0.0
plotly==5.17.0
websocket-client==1.6.4
requests==2.31.0
sqlalchemy==2.0.23
apscheduler==3.10.4
ccxt==4.1.63
pycryptodome==3.19.0
